Update plc parsing and add append wayside method
The plc that is currently being parsed has an error whoops, but I'll figure that out eventually
when we want to figure out the setting up of the waysides, I added a function to add one in the track controller control. Once we parse and split up the track, make an instance of the Wayside controller and fill on the info like in the init of track control controller and append it to the list with the added method.